identical nodes is commented out

+++ src/main/java/group8/algorithm/GreedyHeuristic.java
@@ -11,7 +11,7 @@ import java.util.Map;
 public class GreedyHeuristic implements IHeuristic {
 
     @Override
-    public int calculateEstimate(Schedule state, HashMap<String, Node> allNodes) {
+    public double calculateEstimate(Schedule state, HashMap<String, Node> allNodes) {
         Map<String, int[]> stateNodes = new HashMap<String, int[]>();
         stateNodes.putAll(state.getTasks()); // Get all nodes in the schedule. Put into a new map
         ArrayList<String> stateKeys = new ArrayList<String>(state.getTasks().keySet()); // Keys of the tasks in schedule
